Extract ObjectMappers into separate package extras:jackson-support
