+++ extras/jackson-support/build.gradle
@@ -1,0 +1,11 @@
+apply from: "${rootDir}/gradle/publish.gradle"
+
+dependencies {
+    compile "com.fasterxml.jackson.core:jackson-databind"
+    compile "com.fasterxml.jackson.datatype:jackson-datatype-guava"
+    compile "com.fasterxml.jackson.datatype:jackson-datatype-jdk7"
+    compile "com.fasterxml.jackson.module:jackson-module-afterburner"
+
+    testCompile "junit:junit"
+    testCompile "org.assertj:assertj-core"
+}
