+++ retrofit2-clients/build.gradle
@@ -3,6 +3,7 @@ apply plugin: "org.inferred.processors"
 apply from: "${rootDir}/gradle/publish.gradle"
 
 dependencies {
+    compile project(":extras:jackson-support")
     compile project(":error-handling")
     compile project(":http-clients-api")
     compile project(":service-config")
