+++ retrofit2-clients/src/main/java/com/palantir/remoting1/retrofit2/Retrofit2ClientBuilder.java
@@ -26,6 +26,7 @@ import com.palantir.remoting1.clients.ClientConfig;
 import com.palantir.remoting1.config.service.BasicCredentials;
 import com.palantir.remoting1.config.service.ProxyConfiguration;
 import com.palantir.remoting1.config.ssl.TrustContext;
+import com.palantir.remoting1.ext.jackson.ObjectMappers;
 import com.palantir.remoting1.tracing.okhttp3.OkhttpTraceInterceptor;
 import java.io.IOException;
 import java.util.List;
